@@ -15,15 +15,13 @@ class ProtocolRegistrarImplWindows extends ProtocolRegistrar {
     String appPath = Platform.resolvedExecutable;
 
     String protocolRegKey = 'Software\\Classes\\$scheme';
-    RegistryValue protocolRegValue = const RegistryValue(
+    RegistryValue protocolRegValue = const RegistryValue.string(
       'URL Protocol',
-      RegistryValueType.string,
       '',
     );
     String protocolCmdRegKey = 'shell\\open\\command';
-    RegistryValue protocolCmdRegValue = RegistryValue(
+    RegistryValue protocolCmdRegValue = RegistryValue.string(
       '',
-      RegistryValueType.string,
       '"$appPath" "%1"',
     );
 
